Move EventPipeEventPayload to its own file.
